Added and fixed ITs
