+++ packages/yourloops/test/integration/pages/daily/daily-view.caregiver.spec.tsx
@@ -31,9 +31,9 @@ import { mockDataAPI } from '../../mock/data.api.mock'
 import { mockNotificationAPI } from '../../mock/notification.api.mock'
 import { unmonitoredPatientId } from '../../data/patient.api.data'
 import { mockDirectShareApi } from '../../mock/direct-share.api.mock'
-import { checkPatientNavBarAsCaregiver } from '../../assert/patient-nav-bar'
+import { checkPatientNavBarAsCaregiver } from '../../assert/patient-nav-bar.assert'
 import { renderPage } from '../../utils/render'
-import { checkCaregiverLayout } from '../../assert/layout'
+import { checkCaregiverLayout } from '../../assert/layout.assert'
 import { UserRole } from '../../../../lib/auth/models/enums/user-role.enum'
 import { mockUserApi } from '../../mock/user.api.mock'
 import { mockPatientApiForCaregivers } from '../../mock/patient.api.mock'
