+++ packages/yourloops/test/integration/pages/daily/daily-view.hcp.spec.tsx
@@ -33,7 +33,7 @@ import { mockNotificationAPI } from '../../mock/notification.api.mock'
 import { unmonitoredPatientId } from '../../data/patient.api.data'
 import { mockChatAPI } from '../../mock/chat.api.mock'
 import { mockDirectShareApi } from '../../mock/direct-share.api.mock'
-import { checkPatientNavBarAsHCP } from '../../assert/patient-nav-bar'
+import { checkPatientNavBarAsHCP } from '../../assert/patient-nav-bar.assert'
 import { renderPage } from '../../utils/render'
 import { mockUserApi } from '../../mock/user.api.mock'
 import { mockPatientApiForHcp } from '../../mock/patient.api.mock'
