+++ packages/yourloops/test/integration/pages/notification/notification.patient.spec.tsx
@@ -30,10 +30,10 @@ import { renderPage } from '../../utils/render'
 import { mockPatientLogin } from '../../mock/patient-login.mock'
 import { monitoredPatientAsTeamMember } from '../../data/patient.api.data'
 import NotificationApi from '../../../../lib/notifications/notification.api'
-import { iTeamOne, buildTeamOne, buildTeamTwo } from '../../mock/team.api.mock'
+import { buildTeamOne, buildTeamTwo, iTeamOne } from '../../mock/team.api.mock'
 import TeamAPI from '../../../../lib/team/team.api'
 import { invitationTeam } from '../../data/notification.data'
-import { checkAcceptTeamInvite, closeDialogNotificationTeam } from '../../assert/notification-join-team'
+import { checkAcceptTeamInvite, closeDialogNotificationTeam } from '../../assert/notification-join-team.assert'
 
 describe('Notification page for patient', () => {
   beforeAll(() => {
