+++ packages/yourloops/test/integration/pages/profile/profile.caregiver.spec.tsx
@@ -35,10 +35,10 @@ import {
 import { mockTeamAPI } from '../../mock/team.api.mock'
 import { mockNotificationAPI } from '../../mock/notification.api.mock'
 import { act, fireEvent, screen, waitFor, within } from '@testing-library/react'
-import { checkCaregiverLayout } from '../../assert/layout'
+import { checkCaregiverLayout } from '../../assert/layout.assert'
 import { mockDirectShareApi } from '../../mock/direct-share.api.mock'
 import { mockPatientApiForHcp, mockPatientApiForPatients } from '../../mock/patient.api.mock'
-import { checkCaregiverProfilePage, checkPasswordChangeRequest } from '../../assert/profile'
+import { checkCaregiverProfilePage, checkPasswordChangeRequest } from '../../assert/profile.assert'
 import userEvent from '@testing-library/user-event'
 import { type Profile } from '../../../../lib/auth/models/profile.model'
 import { type Settings } from '../../../../lib/auth/models/settings.model'
