+++ packages/yourloops/test/integration/pages/profile/profile.hcp.spec.tsx
@@ -32,7 +32,7 @@ import { mockNotificationAPI } from '../../mock/notification.api.mock'
 import { act, fireEvent, screen, waitFor, within } from '@testing-library/react'
 import { mockDirectShareApi } from '../../mock/direct-share.api.mock'
 import { mockPatientApiForHcp } from '../../mock/patient.api.mock'
-import { checkHcpProfilePage, checkPasswordChangeRequest } from '../../assert/profile'
+import { checkHcpProfilePage, checkPasswordChangeRequest } from '../../assert/profile.assert'
 import userEvent from '@testing-library/user-event'
 import { type Profile } from '../../../../lib/auth/models/profile.model'
 import { type Settings } from '../../../../lib/auth/models/settings.model'
