+++ packages/yourloops/test/integration/pages/signup-stepper/signup.caregiver.spec.tsx
@@ -32,11 +32,16 @@ import {
   loggedInUserId,
   mockAuth0Hook
 } from '../../mock/auth0.hook.mock'
-import { checkAccountSelectorStep, checkConsentStep, checkProfileStep, checkStepper } from '../../assert/signup-stepper'
+import {
+  checkAccountSelectorStep,
+  checkConsentStep,
+  checkProfileStep,
+  checkStepper
+} from '../../assert/signup-stepper.assert'
 import { mockUserApi } from '../../mock/user.api.mock'
 import userEvent from '@testing-library/user-event'
 import { renderPage } from '../../utils/render'
-import { checkFooterForUserNotLoggedIn } from '../../assert/footer'
+import { checkFooterForUserNotLoggedIn } from '../../assert/footer.assert'
 import { UserRole } from '../../../../lib/auth/models/enums/user-role.enum'
 import { CountryCodes } from '../../../../lib/auth/models/country.model'
