+++ packages/yourloops/test/integration/pages/signup-stepper/signup.hcp.spec.tsx
@@ -32,14 +32,19 @@ import {
   loggedInUserId,
   mockAuth0Hook
 } from '../../mock/auth0.hook.mock'
-import { checkAccountSelectorStep, checkConsentStep, checkProfileStep, checkStepper } from '../../assert/signup-stepper'
+import {
+  checkAccountSelectorStep,
+  checkConsentStep,
+  checkProfileStep,
+  checkStepper
+} from '../../assert/signup-stepper.assert'
 import { mockUserApi } from '../../mock/user.api.mock'
 import { HcpProfession } from '../../../../lib/auth/models/enums/hcp-profession.enum'
 import userEvent from '@testing-library/user-event'
 import { renderPage } from '../../utils/render'
 import { UserRole } from '../../../../lib/auth/models/enums/user-role.enum'
 import { CountryCodes } from '../../../../lib/auth/models/country.model'
-import { checkFooterForUserNotLoggedIn } from '../../assert/footer'
+import { checkFooterForUserNotLoggedIn } from '../../assert/footer.assert'
 
 describe('Signup stepper as hcp', () => {
   const { updateAuth0UserMetadataMock } = mockUserApi()
