+++ packages/yourloops/test/integration/pages/trends/trends-view.caregiver.spec.tsx
@@ -33,9 +33,9 @@ import { mockNotificationAPI } from '../../mock/notification.api.mock'
 import { mockPatientApiForCaregivers } from '../../mock/patient.api.mock'
 import { mockMedicalFilesAPI } from '../../mock/medical-files.api.mock'
 import { mockDirectShareApi } from '../../mock/direct-share.api.mock'
-import { checkPatientNavBarAsCaregiver } from '../../assert/patient-nav-bar'
+import { checkPatientNavBarAsCaregiver } from '../../assert/patient-nav-bar.assert'
 import { renderPage } from '../../utils/render'
-import { checkCaregiverLayout } from '../../assert/layout'
+import { checkCaregiverLayout } from '../../assert/layout.assert'
 import { UserRole } from '../../../../lib/auth/models/enums/user-role.enum'
 import { mockUserApi } from '../../mock/user.api.mock'
 import { unmonitoredPatientId } from '../../data/patient.api.data'
