+++ packages/yourloops/test/integration/pages/verify-email/verify-email.common.spec.tsx
@@ -29,7 +29,7 @@ import * as auth0Mock from '@auth0/auth0-react'
 import { getAccessTokenWithPopupMock, logoutMock } from '../../mock/auth0.hook.mock'
 import { renderPage } from '../../utils/render'
 import { screen, waitFor, within } from '@testing-library/react'
-import { checkFooterForUserNotLoggedIn } from '../../assert/footer'
+import { checkFooterForUserNotLoggedIn } from '../../assert/footer.assert'
 import userEvent from '@testing-library/user-event'
 import { AUTH0_ERROR_EMAIL_NOT_VERIFIED } from '../../../../lib/auth/models/auth0-error.model'
 import { AppRoute } from '../../../../models/enums/routes.enum'
